Auditoria final e preparacao para producao

+++ models/plano_tipo.py
@@ -23,7 +23,8 @@ class PlanoTipo(db.Model):
 
     id = db.Column(db.Integer, primary_key=True)
     nome = db.Column(db.String(50), nullable=False, unique=True)
-    preco = db.Column(db.Integer, nullable=False)
+    # preco em REAIS (Float) — preserva centavos (ex: 109.99)
+    preco = db.Column(db.Float, nullable=False)
     descricao = db.Column(db.Text, nullable=True)
     dias_permitidos = db.Column(db.String(30), nullable=False, default="1,2,3,4")
